@@ -1,18 +1,22 @@
 package org.lamisplus.modules.base.config;
 
 import com.foreach.across.core.annotations.Exposed;
+import io.github.jhipster.config.JHipsterProperties;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @ConfigurationProperties(prefix = "lamis", ignoreUnknownFields = false)
 @Configuration
 @Exposed
 @Getter
 @Setter
-@EnableAutoConfiguration
+@Import({
+    JHipsterProperties.class,
+    org.springframework.boot.autoconfigure.mail.MailSenderAutoConfiguration.class
+})
 public class ApplicationProperties {
     private String modulePath = "modules";
     private String databaseDir;
