@@ -5,14 +5,16 @@ import org.lamisplus.modules.base.domain.entities.Form;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Repository
 public interface FormRepository extends JpaRepository<Form, String> {
-    @Query("select t from Form t join t.module m where t.name = ?1 and m.active = true order by t.priority desc")
+    @Query("select t from Form t join t.module m where t.name = ?1 and m.active = true and m.started = true order by t.priority desc")
     List<Form> findByName(String name, Pageable pageable);
 
     List<Form> findByNameAndModule(String name, Module module);
 
-    List<Form> findByModule_ActiveTrue();
+    List<Form> findByModule_StartedTrue();
 }
